requirements.txt
langchain==0.2.5
langchain-google-genai
streamlit
python-dotenv
tiktoken
pinecone-client
langchain_experimental
tabulate